Revert "Revert "Merge branch 'main' of https://github.com/is212-ljps/is212-LJPS into Lance/ljps-57""
This reverts commit 24acb3db1b52d6e78e50e03e697ecfb4d775402e.

+++ backend/service/src/SkillsService.js
@@ -1,3 +1,5 @@
+const utils = require('../../util')
+
 function skillsService(database){
   const skillsService = {}
 
@@ -18,6 +20,9 @@ function skillsService(database){
   }
 
   skillsService.getSkillById = async (skillId) => {
+    if(!skillId){
+      return false
+    }
     try {
       return await database.getSkillById(skillId)
     } catch (err) {
@@ -34,6 +39,9 @@ function skillsService(database){
   } 
 
   skillsService.getCoursesAssignedToSkill = async (skillId) => {
+    if(!skillId){
+      return false
+    }
     try {
       return await database.getCoursesAssignedToSkill(skillId)
     } catch (err) {
@@ -42,9 +50,17 @@ function skillsService(database){
   }
 
   skillsService.createSkill = async (skillName, skillDescription, assignedCourses) => {
+    if(!skillName || !assignedCourses){
+      return false
+    } else if (utils.checkLength(5,50, skillName) || utils.checkLength(0,300, skillDescription) || utils.checkLength(1,9999, assignedCourses)) {
+      return false
+    }
+
     try {
       const data = await database.createSkill(skillName, skillDescription);
-      await database.assignCoursesToSkills(assignedCourses, data.insertId)
+      if (assignedCourses) {
+        await database.assignCoursesToSkills(assignedCourses, data.insertId)
+      }
       return data.insertId;
     } catch (err) {
       throw err;
@@ -52,6 +68,9 @@ function skillsService(database){
   }
 
   skillsService.deleteSkillById = async (skillId) => {
+    if(!skillId){
+      return false
+    }
     try {
       return await database.deleteSkillById(skillId)
     } catch (err) {
@@ -60,6 +79,11 @@ function skillsService(database){
   }
 
   skillsService.updateSkillById = async (skillId, skillName, skillDescription, assignedCourses) => {
+    if(!skillId || !skillName || !assignedCourses){
+      return false
+    } else if (utils.checkLength(5,50, skillName) || utils.checkLength(0,300, skillDescription) || utils.checkLength(1,9999, assignedCourses)) {
+      return false
+    }
     try {
       const data = await database.updateSkillById(skillId, skillName, skillDescription);
       await database.removeCoursesFromSkill(skillId)
